requirements.txt
Flask==1.1.2
influxdb==5.3.1
pandas==1.2.2
pandas-gbq==0.14.1
google-cloud-storage==1.36.0
google-cloud-bigquery==2.10.0
plotly==4.14.3
dataverk==0.3.24
gcsfs==0.7.2
fiaas-logging==^0.1.1


pytz~=2021.3
Werkzeug~=2.0.2
requests~=2.26.0